fix(profile): Explicitly cast to 32 bit.

+++ dsp/src/kernel/knl_profile.c
@@ -63,8 +63,8 @@ t_profile knl_profile_stats(void) {
 
     static t_profile stats;
 
-    stats.period = sport0_period();
-    stats.cycles = g_module_cycles;
+    stats.period = (uint32_t)sport0_period();
+    stats.cycles = (uint32_t)g_module_cycles;
 
     return stats;
 }
